requirements.txt
fastapi
uvicorn[standard]
openai
streamlit
chromadb
pydantic
pydantic-settings
python-dotenv
Pillow
python-multipart
